refactor(brick-generator): extract shared reference content resolution
Move annotation transforms into reference_content_transforms.dart and
add resolveReferenceContent() so single-file preview can reuse the same
pipeline as full brick generation.
